Added marketplace and changed files
Made the market place and changed the login from the index js to its own file
